Stop the warning gate matching clang's summary line; fix what it then found
Running the new AArch64 warning lane end to end surfaced two things the pass
that added it had not yet seen.

src/c/ama_ed25519.c has never been inside the strict-warnings gate
------------------------------------------------------------------
CMakeLists.txt drops that file in favour of the donna shim whenever
AMA_ED25519_ASSEMBLY is on, which is the x86-64 DEFAULT.  Verified: an
x86-64 build produces zero ama_ed25519.c objects and links
ed25519_donna_shim.c instead.  So the entire in-tree Ed25519 backend — field
arithmetic, point decompression, the verify path, the code every
non-x86-64 target ships — sat outside the one job that examines compiler
diagnostics.  The AArch64 lane is the first configuration to compile it under
the strict flag set.

It carried exactly one finding: `s[31] ^= fe25519_isnegative(x) << 7`
narrowing int to uint8_t (-Wconversion).  fe25519_isnegative returns {0, 1},
so the shifted value is {0, 0x80} and the narrowing is exact — the two forms
are value-equivalent for every input.  Fixed with the explicit cast the Kyber
packers already use.

The cast is value-equivalent but NOT codegen-neutral, and the difference is
worth stating rather than glossing: giving GCC the narrower type changes what
it can prove about the expression, and because ge25519_p3_tobytes is `static`
and inlined into several callers that propagates through inlining and
register allocation across the whole translation unit.  Measured on aarch64
at -O3 -funroll-loops: ama_ed25519.c.o is the ONLY object that changes among
all of them, and linked .text moves 469,124 -> 464,852 bytes.  x86-64 is
untouched (the file is not in that build): the x86-64 library is byte-
identical to 267c16c, SHA-256
2030cbbcade673dc18060dac464dc8237e6f9e32b5d13885a1986de614bf2d82.

Behavioural equivalence is carried by the AArch64 suite, not by assertion:
ctest is 64/64 under QEMU user-mode, including the RFC 8032 Ed25519 KATs and
the donna-vs-fe51 backend differential.

The gate was matching clang's own bookkeeping
---------------------------------------------
_WARNING_RE was `\bwarning[ :]`, which matches `1 warning generated.` —
clang's per-translation-unit SUMMARY line, printed once per file that emitted
any warning at all, INCLUDING warnings this allowlist had already excused.
Twelve bogus findings across the two clang configurations.  GCC prints no
such line, which is why it survived the first runs; the first clang log the
gate ever saw was the one that exposed it.

The matcher now requires the colon that `warning:` and MSVC's
`warning C4244:` share, and declines the plural `2 warnings generated.`
through `\bwarning\b`.  Four cases pin it, including that a summary line
sitting beside a real diagnostic does not mask it.

Verified at this commit:
  strict gcc/clang x noopt/Release, gate over all four logs   exit 0
  strict AArch64 NEON + SVE2, gate over both logs             exit 0
  ctest AArch64 (QEMU user-mode)                              64/64, exit 0
  pytest tests/test_compiler_warning_gate.py                  24 passed, exit 0
  documented counts (62) / ruff / black / mypy --strict       exit 0

+++ src/c/ama_ed25519.c
@@ -229,7 +229,18 @@ static void ge25519_p3_tobytes(uint8_t *s, const ge25519_p3 *h) {
     fe25519_mul(x, h->X, recip);
     fe25519_mul(y, h->Y, recip);
     fe25519_tobytes(s, y);
-    s[31] ^= fe25519_isnegative(x) << 7;
+    /* Explicit cast, in the style the Kyber packers use: fe25519_isnegative
+     * returns int in {0, 1} (it is the low bit of the canonical encoding), so
+     * the shifted value is {0, 0x80} and the narrowing is exact.  Without the
+     * cast the compound assignment converts int -> uint8_t implicitly, which
+     * -Wconversion reports.
+     *
+     * This file is compiled ONLY on non-x86-64 targets — CMakeLists.txt drops
+     * it in favour of the donna shim when AMA_ED25519_ASSEMBLY is on, which is
+     * the x86-64 default — so the whole in-tree Ed25519 backend sat outside
+     * the strict-warnings gate until that gate gained an AArch64 lane.  This
+     * was the one finding it carried. */
+    s[31] ^= (uint8_t)(fe25519_isnegative(x) << 7);
 }
 
 /*
